Finished exercise

+++ src/main/java/com/codeup/springblog/controller/PostController.java
@@ -1,19 +1,16 @@
-
 package com.codeup.springblog.controller;
 
-import com.codeup.springblog.EmailService;
+import com.codeup.springblog.services.EmailService;
 import com.codeup.springblog.model.Post;
 import com.codeup.springblog.model.User;
 import com.codeup.springblog.repositiories.PostRepo;
-import com.codeup.springblog.repositiories.UserRepository;
-import org.hibernate.service.spi.InjectService;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.codeup.springblog.repositiories.UserRepository;;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.thymeleaf.model.IModel;
 
-import java.util.ArrayList;
+import java.security.Principal;
 import java.util.List;
 
 @Controller
@@ -38,25 +35,34 @@ public class PostController {
     }
 
     @GetMapping("/posts/{id}")
-    public String post(Model model, @PathVariable long id) {
+    public String post(Model model, @PathVariable long id, Principal principal) {
         model.addAttribute("title", "View Post");
+        String userName = "";
+        if(principal != null){
+            userName = principal.getName();
+        }
         Post post = postDao.findById(id);
-        User user = userDao.findById(1L);
         model.addAttribute("post", post);
-        model.addAttribute("user", user);
+        model.addAttribute("username", userName);
         return "posts/show";
 
     }
 
     @GetMapping("/posts/create")
     public String showCreateForm(Model model) {
-        model.addAttribute("post", new Post());
-        return "posts/create";
+        User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(loggedInUser != null) {
+            model.addAttribute("post", new Post());
+            return "posts/create";
+        } else{
+            return "redirect:/login";
+        }
     }
 
     @PostMapping("/posts/create")
     public String postNewPost(@RequestParam String title, @RequestParam String body) {
-        Post post = new Post(title, body, userDao.findById(1L));
+        User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Post post = new Post(title, body, loggedInUser);
         postDao.save(post);
         emailService.prepareAndSend(post,"A new post has been created!", "You created a new post!");
         return "redirect:/posts";
@@ -64,15 +70,23 @@ public class PostController {
 
     @PostMapping("/posts/{id}/delete")
     public String delete(@PathVariable long id){
-        postDao.deleteById(id);
+        User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(loggedInUser.getId() == postDao.getOne(id).getUser().getId()) {
+            postDao.deleteById(id);
+        }
         return "redirect:/posts";
     }
 
     @GetMapping("/posts/{id}/edit")
     public String editForm(@PathVariable long id, Model model) {
-        Post postToEdit = postDao.getOne(id);
-        model.addAttribute("post", postToEdit);
-        return "posts/edit";
+        User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (loggedInUser != null) {
+            Post postToEdit = postDao.getOne(id);
+            model.addAttribute("post", postToEdit);
+            return "posts/edit";
+        }else{
+            return "redirect:/posts";
+        }
     }
 
     @PostMapping("/posts/{id}/edit")
